ArduinoJson validation and vs-config

+++ lib/updates/updates.h
@@ -4,7 +4,7 @@
 
 typedef struct fields
 {
-    int delta_t;
+    unsigned long delta_t;
     float field1, field2, field3, field4, field5;
 } fields;
